New animation/effect when selecting a contact.
In preparation for the new design here is a new fade transition + blurry effect (based on @benfrancis recommendation of using an svg filter) when selecting a contact.

For performance purpose we first let the fade animation finish before adding the blurry effect (the css-transition framerate is poor on the device if we animate the opacity of the div with its -moz-element background).

This will of course look way better with the proper design in place, just doing the groundwork :)

+ new test covering this