Adding WiFi settings support for EAP-TLS mode
This patch introduces EAP support with EAP-TLS first. The <select> is
multiple because without you can't pick a choice for now. It provides
TLS, TTLS, PEAP yet only TLS is currently implemented. Specifying
certificates (CA, user, privatekey) means giving the full path of the
corresponding filename in the mobile system. For instance, upload you
certificates in /data/misc/wifi/ with adb push, and put the full path in
the textbox. Filepicker is broken, so we can't use it. Only PKCS#12 is
supported for private key in TLS. Tested with a CA certificate in PEM
format. Anything that is supported by WPA Supplicant should be good.

Later, using a key store for handling certificates (as Android) might be
a good idea.