update display on header changes, fix header disp, show atts
We weren't updating the DOM when our headers changed, we do now.

Abandoned inline-block for header-display purposes because it totally
did not work.

We now show poor man's stars (*) and attachment icons (&) in the
header.  We also properly list attachments in the body display (thanks
to gaia email lib fixes.)