improve inert html reply implementation, make text area grow/shrink
Replying to an HTML message puts the HTML all inert below the textarea
that the user can actually edit.  We reuse the iframe logic that we use
for display, but make it non-interactive so the user doesn't end up in
a hellish scrolling world that sucks.  I tried to use preventDefault
and/or stopPropagation via a modified gesture detector to stop the
automatic scrolling of the page if the embedded HTML page could scroll,
but that didn't work and the user experience sucked.

So now we just zoom out in the newsletter case and make it impossible to
zoom in, but thereby allow a seamless-iframe experience.  (seamless wouldn't
work in this case even if we had it because of the transform.)

The transform may be super-sucky performance-wise and we may need to
create a platform issue, etc.

The textarea now grows/shrinks with the input to avoid being in a
situation where one would need to scroll the contents of the textarea.
This seems desirable in general (to avoid sub-scrolling areas), but is
particularly essential for the combined textarea/read-only html display
to work at all.