[l10n] Sometimes, l10n does not get loaded properly
When the system app starts, in its early days, it sometimes seems not
being able to find the correct navigator.language value. Hence it
defaults to 'en-US' while the rest of the system will be correctly
localized. This has been hit on both Nexus S and Otoro. This patch
mitigates the issue by forcing re-reading the 'language.current'
settings and when receiveing the value, relocalizing the document.