Update test docs in README r=lightsofapollo a=readme-only
The thinking is that MDN is great for longer-explanation docs, but it's
important to have basic documentation in the repository that's versioned
with the code and also available offline.

This adds integration test documentation, adds some basic unit test
documentation, and adds a note about watching files and automatically
running tests.