Bug 821985 - Easying dogfooding backup/reset/restore process
This commit introduces new Makefile targets:
 - backup, which performs the backup of user's data
   BACKUP_DIR (by default $PWD/device-backup).
   The backup operation pull data from indexedDB/ and webapps/.
 - restore, which performs the inverse operation
 - reset-restore, which performs a reset-gaia before restoring
 - reinstall, which does a backup then a reset-restore

The reset-restore can be configured to not restore some databases,
thought the RESTORE_CLEAN variable. By default, activities only will not
be restored, but one can also ask for: settings, sms, netstats,
contacts and alarms. This is useful when dogfooding and debugging, to
recover the full phone data while restarting on good basis for some
parts.