Bug 828174: Track Bluetooth discovery state in Settings app [r=ehung]
The settings app unconditionally stops Bluetooth discovery when
processing incoming events. Stopping discovery when no discovery
is actually running will make the Bluetooth daemon generate an
error.

This patch prevents this by tracking the discovery state in the
settings application and only stopping discovery if it's currently
running. We reuse the value 'discoverTimeout' for this purpose,
because it exposes the required behaviour.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>