Bug 834653 - [EMAIL] Message list synchronization progress bar should become indeterminate/show "candybar" state after a timeout since last progress bar update. v1.0.1 specific building blocks style fix.
Visual design really wants this to look like it does on master.  For risk
reasons, we are manually copying the file into the e-mail repo (and its
images) so that we don't cause any weird fallout in other apps.  This is
basically what the automated build mechanism does, so this isn't a huge deal.

Requested in https://bugzilla.mozilla.org/show_bug.cgi?id=834653#c16