Bug 838486 - [Lockscreen] Lock screen bounce animation not running r=timdream
Before Bug 830480, a callback was called sometimes synchronously and sometimes
asynchronously. In Bug 830480 it was made always asynchronous.

As a result, when switching panels, the "cleaning" code for the unloaded panel
is now always done after the "initializing" code for the loaded panel.

The "main" panel is enabling the bounce animation at initialization and
disabling at unload. Therefore it was disabled right after being enabled.

Note that the bounce animation is also enabled at lock, and disabled at unlock.

This patch skip the loading/unloading command when we ask to switch to the panel
that is already displayed. This should also make the performance better, because
it was actually called several times.
(cherry picked from commit c9617dfaaf0b9d2889c1bdc7b02b471f8f493c0e)