Bug 843511 - [SMS] Send button is not disabled after entering Max Segments in the SMS composer r=borja
Use an alert to display that the max message length for an SMS is reached.
Added tests for the counter text and the alert overlay
(v1.0.1 branch only) Picked the use of this._mozSms instead of navigator.mozSms
so that it could be ultimately testable.

(cherry picked from commit f5a6fc00b3632b79bb2e85a1827a246cb47b311e)