Bug 862704 - Fix failing MobileOperaror unit tests r=alive
The errors really come from a bad status bar test teardown. So this patch is
adding a teardown operation to the mock for MobileOperator, and is fixing the
teardown of the tests for status bar.

Sadly, we were really testing the mock for MobileOperator because of that.

This is also reverting part of Bug 844738.