Bug 863466 - [system] focus the alert/confirm/etc dialogs so that the keyboard is dismissed r=alive,benfrancis
This changes the dialogs in both the browser and the system, for the apps
content and the web content.

In this patch we focus the whole dialog, so that no default action is selected
by default. This effectively fixes the initial bug, and probably makes
accessibility better, even if we'll need more work for this.