Bug 877776 - [MMS] Attach four images cause input text field can't be scrolled properly r=gnarf37
The call to `composeCheck` in `Compose.onAttachClick` is unecessary.
Previously, it was invoked synchronously after
`Compose.requestAttachment` (before any change in the UI could have
occurred). Because `Compose.append` itself invokes `composeCheck`, it is
not necessary to explicitly call it in the `onsuccess` handler for
`Compose.requestAttachment`.

The maximum Compose input height must be updated when transitioning
between pages to account for differences in the height of each page's
header.

Because we must dynamically update the Compose form's height (see the
`ThreadUI.updateInputHeight` method), it is inappropriate to declare a
static `max-height` on that element.

In addition to setting the maximum input height, the "resize" event
handler should recalculate the height of the Compose form. This enables
proper scrolling within the Compose form when the keyboard is
summoned/dismissed.

Finally, in order to follow convention, this commit declares the
"resize" event handler as a method of the `ThreadUI` object and defers
binding until invocation of `TheadUI.init`.
(cherry picked from commit bbfbce4abf7244be25bb02eb8aa710ba41ede036)