Bug 882680: Update discovered Bluetooth devices without names
Sometimes we don't receive the name for a discovered Bluetooth device,
and simply display 'Unknown device' in the Settings app. If the device
later announces itself again, we simply ignore the update.

With this patch, updates for 'unknown' devices are not ignored, but
the list entry in thre Settings app gets replaced. If the device now
supplied a name for itself, the list entry will show the device's
name.