Bug 886238: Retrieve number of unlock retries from IccManager
Gaia apps that allow for the unlocking of the SIM card now retrieve
the number of remaining tries from the IccManager. The interface in
MobileConnection is deprecated and will been removed.

The code is currently wrapped in IccHelper, to ensure compatibility
with mozilla-centry and b2g18. The wrapper code will be removed
when b2g18 compatibility is not needed anymore.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>