Bug 888260 - Notify the user that we can't do anything
When being called by <input type="file">, for example, the 'pick'
activity is in such a state that we can be called, but we will just bail
the user out. Previously, this was done with just an empty confirm
dialog. It's better to notify the user that something was not possible
and notify the activity by canceling it.

Also look at bug 883341.