Bug 883822: [SMS] focus the Compose area when started as an activity - r=gnarf
When the Compose screen is triggered by an activity (e.g. when clicking on a message icon from the call log), recipients are inserted and the onscreen keyboard is triggered. The problem is, the focus is set within a recipient block, not in the input itself: thus, the onscreen keyboard doesn’t disappear when the back button is hit.

A possible fix would have been to focus the recipient input rather than the recipient block, but it’s simpler to focus the Compose area instead (and that’s one less tap for the user). That’s what this simple patch does.

Closes gh-10979
(cherry picked from commit e8ae64883ab17ab8c3ff1fe44b346f8456b315c4)