Bug 899487 - Play a notification when changing sound level
When the user changes the sound level, he might want to be able to have
a preview of what it will actually be like in real life. So let's play
the notification sound he has choosen from settings, as a preview.
We also set the correct mozAudioChannelType to ensure that we are
playing the sound in the right output queue where this volume level has
been set. Since we will make use of 'alarm' channel, we need to update
the manifest accordingly.