Bug 905831 - Use correct default configuration
When the user has an unsupported carrier, we use a default search
engine, so we need to pass the callback this configuration. This commits
fixes the issue by passing the callback the search engine configuration
at [DEFAULT_MCC + DEFAULT_MNC] in the 'data' structure, instead of
pasing (DEFAULT_MCC + DEFAULT_MNC) to callback.