Bug 902296 - Ensure that settings from range value are Numbers
When storing the settings that comes from a range input, we do some
rounding using toFixed(). Unfortunately, toFixed() returns a String.
Changes in the Gecko handling of some of the settings (namely, audio
policy) makes usage of String impossible and we need to ensure that we
pass some Number. We thus change the code to perform a supplementary
parseFloat() after toFixed() to ensure this.