Catch and re-throw error
The wrapped function call is definitely raising an exception, but it
does not appear to be caught by `window.onerror` (which is Mocha's means
of detecting errors in asynchronous test routines).

Catch the error and re-throw an explicit error that shares the same
message in an attempt to allow Mocha to recognize (and report) the
error.