Bug 925184 - [B2G][Settings] The user cannot export contacts to the memory card when USB Storage is enabled
Changed the approach to not to use UMS and settings, but just the sdcard state.

For that added the state 'SHARED' as not available to use for the import/export, since we could have
inconsistencies while using.

Now our utility for sdcard has an extra option to subscribe to changes in the sdcard.