Bug 966243: Fix ICC detection handlers in Settings app, r=arthur.chen
While running the Settings app several errors about self.* not being
defined show up in the logcat. An example is

  E/GeckoConsole(  440): [JavaScript Error: "TypeError: self.updateSimPinsUI \
    is not a function" {file: "app://settings.gaiamobile.org/js/simcard_lock.js" \
    line: 232}]

This is caused by 'self' not being set in the event handlers for ICC
detection. As a result, the UI update does not work correctly when the
ICC changes state. Setting the variable 'self' fixes this problem.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>