Bug 965765 - Listen on simslot-iccinfochange to ensure SIM PIN dialog is displayed
In some cases, we can have 'iccdetected' triggered after
'cardstatechange'. This makes the handler not installed in time and thus
we miss the event and are not able to display the SIM PIN unlock
interface. We add a listener in SimLock object against the
'simlock-iccinfochange' event and check if the SIM card is locked to
display the UI.