Update default.json
Removing http://www.mozilla.org/en-US/firefox/os/.

The content on this website is for desktop and the webdev people have no plans to make it work on mobile. The size of the content is about 1.7MB of image data.  Clearly not something intended for real use. 

Furthermore, it isn't clear that this content would be useful for anyone who already has a FirefoxOS device.