[bug 988747] email body text should not use Fira Mono font
The email body is a <textarea>, which defaults to monospaced; but we should display it with our normal sans-serif font instead.
I also adjusted the font-size to better match other elements on the page, now that it's the same font.
Not sure if a corresponding line-height adjustment would also be appropriate; I have not actually built and tested this change yet.