Bug 949941 - [Camera] Replace all memory-backed Blob instances with disk-backed File instances after saving to disk so "pick" activity consumers don't break. r=djf, a=blocking-b2g=1.3T+ on bug 988749 (e-mail bug for tracking purposes)
This is a manual uplift of https://github.com/mozilla-b2g/gaia/pull/17136 from
bug 949941.  It is blocking-b2g=1.3T+ on bug 988749 which is a bug that solely
exists to track the uplift work in the e-mail app.

Initial uplift work was performed by bkelly and that work reviewed/confirmed by asuth.