Bug 1008859 - [email] on language change, cached card templates not updated. r=jrburke
This is an aggregate of the fixes for bug 1005760 and bug 1034227, but without
the bits of bug 1034227 that back out bug 1005760.  (The backout happened on
trunk/v2.1 because the l10n code gained MutationObserver powers on v2.1 that
they do not have on v1.3t.)

This aggregation is assembled and tested by asuth.