Bug 1026381 - Ensure keeping user NFC setting during operator customization
When the user performs a change to a setting value, the content has no
way to distinguish whether the value retrieved is a default value or a
user defined value. We circumvent this limitation by switching the falsy
default to a null value, allowing the operatorvariant app to be able to
distinguish between a user defined 'false' and a default 'null'.