Bug 1044297 - Update mem perf tests to use new moz-app-loaded event.
After moz-app-loaded event, application has displayed content and ready
for interation. At this point, we should check for memory consumption
regressions.

Apps that implement new startup events don't need to run startup_tests
test. We check if the running app is not in the white list, if it does,
we skip the test.