Bug 1093093 - Try to provide better error reporting r=janx
When the LogShake feature reports an error, we will try to do as much as
we can to handle it gracefully and give more feedback to the user. One
case that we start with is handling the errno 30 (ro-mounted filesystem)
that happens when the user try to dump while his device is hooked to a
computer and the SD card is shared via UMS. For now, this is the only
one that we handle, so we display a more generic error message for
others.