Bug 1103559 - Voicemail, Captive Portal and Roaming notifications should not be resent r=etienne
We do not want those notifications to be resent, since they depend on
external conditions:
 - Voicemail will be anyway retriggered by RIL code when connecting to
   network,
 - Captive Portal will be redetected when reconnecting to the network,
   which may have changed in-between,
 - Roaming status will also be redetected when rebooting
Gecko offers a mozbehavior object which can contain a noresend property.
We set this for all those notifications. We need, however, to keep the
cleanup code for migration/upgrade usecase: pre-existing sent
notification will still need to be cleanup if we don't want them to be
resent.