Bug 1103560 - Unique system message handler for Notifications r=alive,mhenretty,ggp
The system message API exposed by Gecko only allows to have one handler
per application ; yet, system app do have multiple modules using the
Notification API. If those needs to handle the case of notification
being resent, then they will need to have a system message handle for
the 'notification' message. This is what 'Screenshot' and 'FindMyDevice'
did, which means only one could work, depending on which one gets
executed.

We solve this by introducing a new system service
'handleSystemMessageNotification' that each module should
|Service.request| and exposing its module name. Then a
'handleSystemMessageNotification' will get called on the module when a
matching system message is being received. We leave the logic of what do
to with the notification to the module itself: close it, keep it.