Bug 1116087 - Safely clone elements for HTML cookie cache
Bug 1081039 fixed the platform to call the custom element callback
appropriately when cloning / importing nodes.  A document context
with a fresh registry is required to do this (actually) correctly.

Without this fix the email app's HTML cookie cache would end up
with a corrupted cookie cache state and potentially other app
problems since the rest of the card infrastructure really was not
expecting this type of thing to happen.