Bugs 1112131, 1124582, 1127857, 1129850, 1129943 - Use gaia-header 0.6.x r=julien,wilsonpage,jlorenzo
This changes how gaia-header calculates the font size to squeeze the header text
in the available space. We don't use offsetLeft but instead we calculate the
various buttons widths. In the common case where we only have the action button
we don't even take the current width and assume it's 50 px, which eliminates a
possible reflow.

Additionally we now support the start, end, and no-font-fit arguments.

start, end: allows the component user to gives the left/right button sizes to
the component, which effectively eliminates all reflows.
no-font-fit: when this attribute is present we skip all font resizing algorithm
and thus no reflow. This allows to defer some initialization (and reflow!) for the headers
that are not displayed at startup.

Last but not least now we try to squeeze the font fit for centered titles before
uncentering them.