Bug 1119294 - Add logshake ID to keep multiple notifications around r=janx
Users may want to keep multiple logshake notifications around, so lets
generate a tag that is dependant on the timestamp to provide this. We
use the current time when receiving the capture start event, and keep it
until we receive either a success or a failure event, at which point we
reset it to null.