Bug 1168396 - Adjust Marionette JS client to protocol changes; r=aus
Bug 1153822 to Gecko reduces the amount of data sent will make requests
faster. Aligning the protocol closer to what WebDriver expects, will
reduce the amount of post-processing required in the httpd.

The Marionette protocol version 1 value response looked like this:

	{"from":"0","value":null,"status":0,"sessionId":"{6b6d68d2-4ac9-4308-9f07-d2e72519c407}"}

And this for ok responses:

	{"from":"0","ok":true}

And this for errors:

	{"from":"0","status":21,"sessionId":"{6b6d68d2-4ac9-4308-9f07-d2e72519c407}","error":{"message":"Error loading page, timed out (onDOMContentLoaded)","stacktrace":null,"status":21}}

Bug 1153822 (protocol 2) drops the "from" and "sessionId" fields and
"status" field from non-error responses. It also drops the "ok" field
in non-value responses and flattes the error response.

This patch adds support for protocol version 2 with backwards
compatibility for protocol 1.