Bug 1183829 - Add update script for locale repositories.
Automate some of the instructions for building multilocale gaia.
https://developer.mozilla.org/en-US/Firefox_OS/Building#Building_multilocale

Takes a language.json file as an argument, or uses languages_dev.json
if no argument is given. Then for each locale in the json file it
either clones the corresponding hg repository, or if the directory
already exists tries to update it.