Adding "Power button locks immediately" option (bug 1186124)
This adds a lockscreen settings option to immediately require the
pass code (if configured) after pushing the power button, regardless
of the pass code timeout setting.

This allows to have a long pass code timout as fallback for cases
where the user forgets to manually lock, while giving the user a
choice to leave the device unlocked for long periods of time, for
example during development.