Bug 1209415 - [Contacts] RTL-friendly headers Bug 1209415 - [Contacts] CSS refactoring
Most rules have been made symmetrical. To accomodate the A-Z scrollbar, which is
always on the right edge of the screen, three main changes have been made:

* padding of #view-contacts-list [data-type="list"]
* margin of [data-type="list"] li aside.pack-end
* padding of .group-section (rule has been completely removed)

The has no effect on the LTR view -- to be sharp it fixes a minor issue with the
horizontal separator length -- and it makes the RTL view symmetrical, except
for the A-Z scrollbar.

Conflicts:
	apps/communications/contacts/elements/form.html
	build/csslint/xfail.list