Bug 1210305 - Lock screen implementation
Lock Screen UI implementation as per
http://r1l8an.axshare.com/#p=2_0_lock_screen

Modified Javascript files :-
- notifications.js
- lockscreen_notifications.js
As per the UX, similar kind of notifications should be clubbed together.
Tested by installing an external app and generated local notifications.

- lockscreen_inputpad.js
Password entered by the user on the lockscreen should be visible for the 1
second before changing it to dot.

- lockscreen_state_manager.js
If user clicks on the lockscreen notification and clicks on open after
unlocking the device , the corresponding app should be launched.