Bug 1218547 - [E-mail] fix caret in settings view
Commit a05b061a05fb9e2f624a66258df8ca6125089dac removes an image that we
used: forward-light.svg, thus breaking the caret.

This commit does not put it back, but add a left-light.svg and a
right-light.svg instead, to also fix the caret in RTL. To be consistent
with gaia-icons, these 2 icons have been copied from it, changing just
the color and opacity. That way if we migrate to gaia-icons, less work
will have to be done. right-light seemed to have exactly the same shape
as forward-light anyway.