Bug 1210201 - Work around the better event fluffing algorithm r=jrburke r=jlorenzo
When we get a DOM event for an existing event "<a>" link, maybe it's the result
of Gecko's event retargeting algorithm: the user may have actually tapped around
the link but not right on the link.

While this is generally a good thing, this is not in this case, as it prevents
the user from creating a new event just before or after an existing event. So we
validate this event to see whether it happened really on the button.

Note that for the "+" buttons, we actually keep the retargeting algorithm.