Bug 1210304 - Unused settings has been removed
David's comment:
"Just removing this one select element leave us with an empty panel, doesn't it? I think you should probably remove this entire file and the references to it, right?
And of course you should also remove the homescreen code that reads this grid.layout.cols setting, whereever that code is."
Removed homescreen setting for changing number of columns