Bug 1210304 - David's comment about number of columns in JSON
David's comment:
"Given that the number of columns will not be configurable, I think that it doesn't need to be in this JSON file anymore. Maybe find the code that reads this file and just define a const for the number of columns."
Removed number of columns from default-homescreens.json. Define a const for the number of columns in configurator.js