Bug 1220311 - Use webapps-manage so that IAC works correctly. r=kgrandon, a=mahe
When handling an IAC message, the home screen will be launched as a
regular app and not as a home screen. Because of this, the
homescreen-webapps-manage permission isn't granted and the app fails to
launch with an exception caused by this. Revert to using the old
webapps-manage permission for now.