Bug 1210305 - Lockscreen: Implemented the lockscreen as per the UX.
Lock Screen UI implementation as per
http://r1l8an.axshare.com/#p=2_0_lock_screen

Modified Javascript files :-
- notifications.js
- lockscreen_notifications.js
As per the UX, similar kind of notifications should be clubbed together.
Tested by installing an external app and generated local notifications.

- lockscreen_inputpad.js
Password entered by the user on the lockscreen should be visible for the 1
second before changing it to dot.

- lockscreen_state_manager.js
If user clicks on the lockscreen notification and clicks on open after
unlocking the device , the corresponding app should be launched.

- Added visual design changes.
- Fixed the issues related to David's comment.
- Fixed the issue in passcode screen
- Fixed the issues related to Egor's comment.
- Changed the test cases as per the implementation
- As per the discussion with Egor, in case of lockscreen is on and
  passcode is off from settings, device will be unlocked from any key.
- Fixed the issue in lockscreen in which passcode screen is coming on
  the homescreen.
- Fixed the lint check errors.Errors such as variable used out of scope, variable is already defined are
not fixed as these are not real errors.