Bug 1227991 - client.js sends wrong variable to Marionette server; r=aus
This patch rectifies an obvious mistake in an earlier rebase where the
Marionette client is passing on the unserialised body to the Marionette
server instead of the serialised version stored in the data variable.

This only happens to work right now because the server is not using
protocol level 3 yet.