Bug 1234063 - Always specify sandbox for scripts to be executed; r=mhenretty
All calls to the remote commands executeScript, executeAsyncScript,
and executeJSScript will send the `sandbox` parameter that implicitly
indicates to the Marionette server that the script must be evaluated
inside a sandbox where changes to the global environment _does not_
have any lasting side-effects.

This enables bug 1123506 to implement support for the WebDriver conformant
approach.